+++ A5/Hitbox.h
@@ -3,21 +3,21 @@
 class Hitbox
 {
 public:
-	Hitbox(Entity* et, int dam, DirectX::XMFLOAT3 launch, int start, int active, int end);
+	Hitbox(Entity* et, int dam, DirectX::XMFLOAT3 launch, float start, float active, float end);
 	~Hitbox();
 	Entity* GetEntity();
 	int Damage();
 	DirectX::XMFLOAT3 Launch();
 	DirectX::XMFLOAT3 P2Launch();
-	int Start(); 
-	int Active();
-	int End();
+	float Start(); 
+	float Active();
+	float End();
 private:
 	Entity* entity;
 	int damage;
 	DirectX::XMFLOAT3 launchVector;
-	int startFrames;
-	int activeFrames;
-	int endFrames;
+	float startFrames;
+	float activeFrames;
+	float endFrames;
 };
